test: added module
